FlaskTemplate/{{cookiecutter.project}}/requirements.txt
cookiecutter
flask
flask-mail
flask-script
flask-wtf
flask-sqlalchemy
flask-hashing
flask-login
pyjwt